+++ src/raiplaysound_cli/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["__version__"]
 
-__version__ = "2.2.2"
+__version__ = "2.3.0"